requirements.txt
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.11
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tenacity==8.3.0
urllib3==2.6.3
